캔들 차트 데이터 프레임 일부 수정
